# Improve performance of trace id generation