@ankon 2024-10-23
I'm actually wondering whether we need the seeding at all. It is used for tests (fine), and otherwise just given the current timestamp. 

If that could be removed, one could start looking at other approaches here, for instance a pool of rand's, or a ringbuffer that gets filled independently and the generation just pulling out some bits ...